refactor mlp
